requirements.txt
typer==0.9.0
rich==13.7.0
tree-sitter==0.20.4
tree-sitter-languages==1.10.2
networkx==3.2.1
graphviz==0.20.1
sentence-transformers
scikit-learn
openai
python-dotenv
google-generativeai